requirements.txt
requests>=2.31.0
feedparser>=6.0.10
groq>=0.4.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
schedule>=1.2.0
openpyxl>=3.1.0  # Optional: For Excel export with text wrapping
